Remove missleading statistics
The statistics about usage (descrepancy, and health) would be really
interesting when done right. But there is a level of overhead in the
early stages of usage that I haven't yet put enough time into
understanding. This results in things looking very unhealthy during
low usage, and those same conditions looking absolutely fine with high
usage. At this point, if I kept them in, I would have to introduce
voodoo-style statistics that would be wrong, and more misleading than
helpful.

It's a shame, but I feel that it's the right decision.